requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
flask>=3.0.0
APScheduler>=3.10.0
